Se volvieron a generar los POCOS

+++ Models/Detallecompra.cs
@@ -14,7 +14,7 @@ namespace SGLibreria.Models
         public int IdCompra { get; set; }
         public int IdProducto { get; set; }
         public int Cantidad { get; set; }
-        public float PrecioUnitario { get; set; }
+        public decimal PrecioUnitario { get; set; }
 
         public virtual Compra IdCompraNavigation { get; set; }
         public virtual Producto IdProductoNavigation { get; set; }
